Update our instructions in README.md
Our FLASK application is creating a COM object each time a route is called
when the route thread closes, the COM object is destroyed.
If that is the last instance of VibrationVIEW the application is also closed.
Sending a restore to the window through the interface would prevent this, but simple solution is to execute VibrationVIEW from the script that starts the server.
Additional optimization will allow a single object for the life the the FLASK application.